client/requirements.txt
mss>=9.0.0
pynput>=1.7.6
requests>=2.31.0
websocket-client>=1.7.0
pyperclip>=1.8.2
plyer>=2.1.0
Pillow>=10.2.0
python-dotenv>=1.0.0
pyobjus>=1.2.0; sys_platform == 'darwin'
